CNN/requirements.txt
torch>=2.2
torchvision>=0.17
pandas>=2.0
scikit-learn>=1.3
numpy>=1.24
pillow>=10.0
tqdm>=4.66
pyyaml>=6.0
